requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.6
django-ipware==4.0.2
django-session-timeout==0.1.0
idna==3.4
mailjet-rest==1.3.4
Pillow==9.2.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
